backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic-settings
python-dotenv
pydantic
